refactor routes pages
