+++ src/components/pages/routes/filter-routes/components/filter-hours/hours-picker-accordion/icons/collapseIcon.tsx
