+++ src/components/pages/routes/filter-routes/components/filter-hours/hours-picker-slider.tsx/hoursPickerSlider.tsx
