+++ src/components/pages/routes/filter-routes/components/filter-hours/icons/backwardIcon.tsx
