+++ src/components/pages/routes/filter-routes/components/filter-price/filterPrice.module.sass
