+++ src/components/pages/routes/filter-routes/components/filter-price/price-range/priceRange.module.sass
