+++ src/components/pages/routes/filter-routes/filterRoutes.tsx
@@ -7,10 +7,8 @@ import FilterHoursEnd from "./components/filter-hours/filterHoursEnd";
 import FilterHoursStart from "./components/filter-hours/filterHoursStart";
 import FilterPrice from "./components/filter-price/filterPrice";
 import s from "./filterRoutes.module.sass";
-import { useRedirectOnSettingsChanged } from "./hooks/useRedirectOnSettingsChanged";
 
 const FilterRoutes = () => {
-  useRedirectOnSettingsChanged();
 
   return (
     <div className={s.container}>
