+++ src/components/pages/routes/routes-block/components/routes-item/components/train-car-classes/trainCarClasses.map.ts
