+++ src/components/pages/routes/routes-block/components/routes-item/components/train/icons/arrowLeftIcon.tsx
