+++ src/components/pages/routes/routes-block/components/routes-item/components/train/icons/arrowRightIcon.tsx
