+++ src/components/pages/routes/routes-block/components/routes-item/components/train/train.module.sass
