+++ src/components/pages/routes/routes-block/components/routes-item/icons/trainIcon.tsx
