+++ src/hooks/useRedirectOnSettingsChanged.ts
