+++ src/pages/routes/index.tsx
@@ -1,18 +1,19 @@
 import React from "react";
 
 import Layout from "@/components/pages/layout";
-import RoutesView from "@/components/pages/routes-view/routesView";
+import RoutesBlock from "@/components/pages/routes/routes-block/routesBlock";
+import RoutesView from "@/components/pages/routes/routes-view/routesView";
 import { useInitDataFromQuery } from "@/hooks/useInitDataFromQuery";
 
 const Routes = () => {
   useInitDataFromQuery();
 
   return (
-    <>
-      <Layout>
-        <RoutesView />
-      </Layout>
-    </>
+    <Layout>
+      <RoutesView>
+        <RoutesBlock />
+      </RoutesView>
+    </Layout>
   );
 };
